wip: modifying api

+++ src/app/api/invitation/route.ts
@@ -34,17 +34,13 @@ export const POST = async (req: NextRequest, res: NextApiResponse<Data>) => {
     // contents,
     custom_url,
     date,
-    post_number,
     primary_image,
     title,
     user_id,
   } = jsonRequest || {}
 
   if (!title || !user_id) {
-    return NextResponse.json(
-      { message: '필수 항목이 없습니다.' },
-      { status: 400 },
-    )
+    return NextResponse.json({ message: '필수 항목이 없습니다.' }, { status: 400 })
   }
 
   const uuid = randomUUID()
@@ -54,256 +50,9 @@ export const POST = async (req: NextRequest, res: NextApiResponse<Data>) => {
       uuid: string
     }[] = []
 
-    const contentPromises = JSON.parse(jsonRequest.contents).map(
-      async (item: any) => {
-        console.log(1111)
+    const contentPromises = JSON.parse(jsonRequest.contents).map(async (item: ContentDataType) => {
+      console.log(1111)
 
-        if (item.type === 'image') {
-          const { isImageUrlsValid } = judgeImageAndVideoValid({
-            image_urls: item.urls,
-          })
-          if (!isImageUrlsValid) {
-            return NextResponse.json(
-              {
-                message:
-                  '적절한 이미지 url이 아닙니다. 공백없이 쉼표로 나누어서 문자열로 보내주세요.',
-              },
-              { status: 400 },
-            )
-          }
-
-          const { data: imageData, error: imageError } = await supabase
-            .from('images')
-            .insert({
-              image_url: item.urls,
-              layout: item.layout,
-              ratio: item.ratio,
-            })
-            .select('id')
-            .single()
-
-          console.log(imageData, 'hmmm????')
-          if (imageError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '이미지 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-
-          if (imageData.id) {
-            refinedContents.push({
-              type: 'images',
-              uuid: imageData.id,
-            })
-          }
-          if (imageError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '이미지 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-        } else if (item.type === 'video') {
-          const { isVideoUrlValid } = judgeImageAndVideoValid({
-            video_url: item.urls,
-          })
-          if (!isVideoUrlValid) {
-            return NextResponse.json(
-              {
-                message:
-                  '적절한 이미지 url이 아닙니다. 공백없이 쉼표로 나누어서 문자열로 보내주세요.',
-              },
-              { status: 400 },
-            )
-          }
-          const { data: videoData, error: videoError } = await supabase
-            .from('videos')
-            .insert({
-              video_url: item.urls,
-              ratio: item.ratio,
-            })
-            .select('id')
-            .single()
-
-          if (videoError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '비디오 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-
-          if (videoData.id) {
-            refinedContents.push({
-              type: 'videos',
-              uuid: videoData.id,
-            })
-          }
-        } else if (item.type === 'text') {
-          const { data: textData, error: textError } = await supabase
-            .from('text')
-            .insert({
-              text: item.text,
-              font_size: item.font_size,
-              font_type: item.font_type,
-              layout: item.layout,
-            })
-            .select('id')
-            .single()
-
-          if (textError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '텍스트 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-
-          if (textData.id) {
-            refinedContents.push({
-              type: 'text',
-              uuid: textData.id,
-            })
-          }
-        } else if (item.type === 'interval') {
-          const { data: intervalData, error: intervalError } = await supabase
-            .from('interval')
-            .insert({
-              size: item.size,
-            })
-            .select('id')
-            .single()
-
-          if (intervalError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '간격 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-
-          if (intervalData.id) {
-            refinedContents.push({
-              type: 'interval',
-              uuid: intervalData.id,
-            })
-          }
-        } else if (item.type === 'map') {
-          const { data: mapData, error: mapError } = await supabase
-            .from('map')
-            .insert({
-              main_address: item.main_address,
-              detail_address: item.detail_address,
-              post_number: item.post_number,
-            })
-            .select('id')
-            .single()
-
-          if (mapError) {
-            return NextResponse.json(
-              {
-                success: false,
-                error: '지도 데이터 삽입 중 오류가 발생했습니다.',
-              },
-              { status: 500 },
-            )
-          }
-
-          if (mapData.id) {
-            refinedContents.push({
-              type: 'map',
-              uuid: mapData.id,
-            })
-          }
-        }
-        return refinedContents
-      },
-    )
-
-    // Promise.all을 사용하여 모든 콘텐츠 처리가 완료될 때까지 기다립니다.
-    refinedContents = (await Promise.all(contentPromises)).filter(Boolean)
-
-    console.log(2222, refinedContents)
-
-    // invitation 테이블에 데이터 삽입
-    const { data, error } = await supabase
-      .from('invitation')
-      .insert([
-        {
-          background_image,
-          contents: refinedContents,
-          custom_url,
-          date,
-          id: uuid,
-          post_number,
-          primary_image,
-          title,
-          user_id,
-        },
-      ])
-      .select('id')
-      .single()
-
-    console.log(333, data, error)
-
-    if (error || !data) {
-      return NextResponse.json(error, { status: 500 })
-    }
-    return NextResponse.json(
-      { id: uuid, message: 'Successfully added' },
-      { status: 200 },
-    )
-  } catch (error) {
-    return NextResponse.json(
-      { message: '초대장 생성 중 오류가 발생했습니다.', error: error },
-      { status: 500 },
-    )
-  }
-}
-
-export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
-  // req의 body에서 값을 받아오도록 변경
-  const jsonRequest = await req.json()
-
-  // next.js14에서는 body에서 데이터를 받아옵니다.
-  const {
-    background_image,
-    contents,
-    custom_url,
-    date,
-    post_number,
-    primary_image,
-    title,
-    user_id,
-    id,
-  } = jsonRequest || {}
-
-  if (!id || !user_id) {
-    return NextResponse.json(
-      { message: 'id field is missing' },
-      { status: 400 },
-    )
-  }
-
-  const uuid = randomUUID()
-
-  let refinedContents: {
-    type: 'images' | 'videos' | 'text' | 'interval' | 'map'
-    uuid: string
-  }[] = []
-  const contentPromises = JSON.parse(jsonRequest.contents).map(
-    async (item: any) => {
       if (item.type === 'image') {
         const { isImageUrlsValid } = judgeImageAndVideoValid({
           image_urls: item.urls,
@@ -328,6 +77,7 @@ export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
           .select('id')
           .single()
 
+        console.log(imageData, 'hmmm????')
         if (imageError) {
           return NextResponse.json(
             {
@@ -423,7 +173,6 @@ export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
         const { data: intervalData, error: intervalError } = await supabase
           .from('interval')
           .insert({
-            invitation_id: uuid,
             size: item.size,
           })
           .select('id')
@@ -473,8 +222,233 @@ export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
           })
         }
       }
-    },
-  )
+      return refinedContents
+    })
+
+    // Promise.all을 사용하여 모든 콘텐츠 처리가 완료될 때까지 기다립니다.
+    refinedContents = (await Promise.all(contentPromises)).filter(Boolean)
+
+    console.log(2222, refinedContents)
+
+    // invitation 테이블에 데이터 삽입
+    const { data, error } = await supabase
+      .from('invitation')
+      .insert([
+        {
+          background_image,
+          contents: refinedContents,
+          custom_url,
+          date,
+          id: uuid,
+          primary_image,
+          title,
+          user_id,
+        },
+      ])
+      .select('id')
+      .single()
+
+    console.log(333, data, error)
+
+    if (error || !data) {
+      return NextResponse.json(error, { status: 500 })
+    }
+    return NextResponse.json({ id: uuid, message: 'Successfully added' }, { status: 200 })
+  } catch (error) {
+    return NextResponse.json(
+      { message: '초대장 생성 중 오류가 발생했습니다.', error: error },
+      { status: 500 },
+    )
+  }
+}
+
+export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
+  // req의 body에서 값을 받아오도록 변경
+  const jsonRequest = await req.json()
+
+  // next.js14에서는 body에서 데이터를 받아옵니다.
+  const { background_image, contents, custom_url, date, primary_image, title, user_id, id } =
+    jsonRequest || {}
+
+  if (!id || !user_id) {
+    return NextResponse.json({ message: 'id field is missing' }, { status: 400 })
+  }
+
+  const uuid = randomUUID()
+
+  let refinedContents: {
+    type: 'images' | 'videos' | 'text' | 'interval' | 'map'
+    uuid: string
+  }[] = []
+  const contentPromises = JSON.parse(jsonRequest.contents).map(async (item: any) => {
+    if (item.type === 'image') {
+      const { isImageUrlsValid } = judgeImageAndVideoValid({
+        image_urls: item.urls,
+      })
+      if (!isImageUrlsValid) {
+        return NextResponse.json(
+          {
+            message: '적절한 이미지 url이 아닙니다. 공백없이 쉼표로 나누어서 문자열로 보내주세요.',
+          },
+          { status: 400 },
+        )
+      }
+
+      const { data: imageData, error: imageError } = await supabase
+        .from('images')
+        .insert({
+          image_url: item.urls,
+          layout: item.layout,
+          ratio: item.ratio,
+        })
+        .select('id')
+        .single()
+
+      if (imageError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '이미지 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+
+      if (imageData.id) {
+        refinedContents.push({
+          type: 'images',
+          uuid: imageData.id,
+        })
+      }
+      if (imageError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '이미지 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+    } else if (item.type === 'video') {
+      const { isVideoUrlValid } = judgeImageAndVideoValid({
+        video_url: item.urls,
+      })
+      if (!isVideoUrlValid) {
+        return NextResponse.json(
+          {
+            message: '적절한 이미지 url이 아닙니다. 공백없이 쉼표로 나누어서 문자열로 보내주세요.',
+          },
+          { status: 400 },
+        )
+      }
+      const { data: videoData, error: videoError } = await supabase
+        .from('videos')
+        .insert({
+          video_url: item.urls,
+          ratio: item.ratio,
+        })
+        .select('id')
+        .single()
+
+      if (videoError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '비디오 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+
+      if (videoData.id) {
+        refinedContents.push({
+          type: 'videos',
+          uuid: videoData.id,
+        })
+      }
+    } else if (item.type === 'text') {
+      const { data: textData, error: textError } = await supabase
+        .from('text')
+        .insert({
+          text: item.text,
+          font_size: item.font_size,
+          font_type: item.font_type,
+          layout: item.layout,
+        })
+        .select('id')
+        .single()
+
+      if (textError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '텍스트 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+
+      if (textData.id) {
+        refinedContents.push({
+          type: 'text',
+          uuid: textData.id,
+        })
+      }
+    } else if (item.type === 'interval') {
+      const { data: intervalData, error: intervalError } = await supabase
+        .from('interval')
+        .insert({
+          invitation_id: uuid,
+          size: item.size,
+        })
+        .select('id')
+        .single()
+
+      if (intervalError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '간격 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+
+      if (intervalData.id) {
+        refinedContents.push({
+          type: 'interval',
+          uuid: intervalData.id,
+        })
+      }
+    } else if (item.type === 'map') {
+      const { data: mapData, error: mapError } = await supabase
+        .from('map')
+        .insert({
+          main_address: item.main_address,
+          detail_address: item.detail_address,
+          post_number: item.post_number,
+        })
+        .select('id')
+        .single()
+
+      if (mapError) {
+        return NextResponse.json(
+          {
+            success: false,
+            error: '지도 데이터 삽입 중 오류가 발생했습니다.',
+          },
+          { status: 500 },
+        )
+      }
+
+      if (mapData.id) {
+        refinedContents.push({
+          type: 'map',
+          uuid: mapData.id,
+        })
+      }
+    }
+  })
   refinedContents = (await Promise.all(contentPromises)).filter(Boolean)
 
   const { data, error } = await supabase
@@ -487,7 +461,6 @@ export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
       custom_url: custom_url ?? jsonRequest.custom_url,
       date: date ?? jsonRequest.date,
       id: uuid,
-      post_number: post_number ?? jsonRequest.post_number,
       primary_image: primary_image ?? jsonRequest.primary_image,
       title: title ?? jsonRequest.title,
       user_id: user_id ?? jsonRequest.user_id,
@@ -497,10 +470,7 @@ export const PUT = async (req: NextRequest, res: NextResponse<Data>) => {
   if (error || !data) {
     return NextResponse.json(error, { status: 500 })
   }
-  return NextResponse.json(
-    { id: uuid, message: 'Successfully added' },
-    { status: 200 },
-  )
+  return NextResponse.json({ id: uuid, message: 'Successfully added' }, { status: 200 })
 }
 
 export const GET = async (req: NextRequest, res: NextApiResponse<Data>) => {
@@ -533,11 +503,7 @@ export const GET = async (req: NextRequest, res: NextApiResponse<Data>) => {
   const { data, error } = await queryBuilder.single()
 
   let contents = []
-  if (
-    data &&
-    data?.contents &&
-    JSON.parse(data?.contents as string).length > 0
-  ) {
+  if (data && data?.contents && JSON.parse(data?.contents as string).length > 0) {
     contents = await Promise.all(
       JSON.parse(data?.contents as string)?.map(async (content: any) => {
         let contentData
@@ -607,18 +573,11 @@ export const DELETE = async (req: NextRequest, res: NextResponse<Data>) => {
   const { id, user_id } = query!
 
   if (!id || !user_id) {
-    return NextResponse.json(
-      { message: 'id와 user_id 필드가 누락되었습니다' },
-      { status: 400 },
-    )
+    return NextResponse.json({ message: 'id와 user_id 필드가 누락되었습니다' }, { status: 400 })
   }
 
   // invitation 테이블에서 데이터 삭제
-  const { error } = await supabase
-    .from('invitation')
-    .delete()
-    .eq('id', id)
-    .eq('user_id', user_id)
+  const { error } = await supabase.from('invitation').delete().eq('id', id).eq('user_id', user_id)
 
   if (error) {
     return NextResponse.json(
@@ -627,8 +586,5 @@ export const DELETE = async (req: NextRequest, res: NextResponse<Data>) => {
     )
   }
 
-  return NextResponse.json(
-    { message: '초대장이 성공적으로 삭제되었습니다' },
-    { status: 200 },
-  )
+  return NextResponse.json({ message: '초대장이 성공적으로 삭제되었습니다' }, { status: 200 })
 }
